Project Two
This is Project Two with brief descriptions of how the pictures and database were completed.  It also has a document stating how some of the difficulties were solved during the process and speaks on MongoDB in its purpose in how it was a model component of the development.  Finally, the document also speaks on how the Dash Framework provides structure to the web application.